Support TikZ externalization library usage
The TikZ Externalization Library provides a high-level automatic or
semi–automatic export feature for TikZ pictures. Its purpose is to
convert each picture to a separate pdf without changing the document
as such. See the tikz/pgfmanual for more details.

The HTWDD logo is one of the TikZ pictures which are able to be
externalized. Thus, it was important to render the HTWDD logo only
once. A customized logo can be used as well. The logo is rendered
at the beginning of the document: \begin{document}.

To use the support of externalization library, the user must activate
the LaTeX shell escape (\write18) feature by using the command-line
option: For TeXLive, this is "-shell-escape"; and for MiKTeX,
it is "-enable-write18". The support is tested with PDFLaTeX and
LuaLaTeX.